requirements.txt
aiofiles==23.1.0
anyio==3.6.2
boto3==1.26.120
botocore==1.29.120
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
fastapi==0.95.1
filelock==3.12.0
fonttools==4.39.3
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy
opencv-python-headless
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.3.0
pydantic==1.10.7
PyMySQL==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.22.0
watchfiles==0.19.0
websockets==11.0.2
IPython==8.13.0
psutil==5.9.5
